requirements.txt
# Web Framework
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Mail==0.10.0
Flask-Limiter==3.7.0
Flask-Babel==4.0.0
WTForms==3.1.2
Werkzeug==3.0.3
gunicorn==22.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35

# Security
cryptography==43.0.1
bcrypt==4.2.0

# Task Queue
celery==5.4.0
redis==5.1.1
flower==2.0.1

# Binance & Trading
python-binance==1.0.19
pandas==2.2.3
numpy==1.26.4

# Historical Data (Yahoo Finance — works from DE, bypasses Binance geo-block)
yfinance>=1.2.0

# Visualization
plotly==5.24.1

# Telegram Bot
python-telegram-bot==21.6

# Payments
stripe==11.1.0

# In-process bot scheduler (replaces Celery for single-container deployments)
APScheduler==3.10.4

# Utils
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.2
Pillow==11.0.0
